添加自动生成代码

+++ src/main/java/com/longbig/multifunction/up/controller/SubjectController.java
@@ -26,7 +26,7 @@ public class SubjectController {
     * @return 单条数据
     */
     @GetMapping("selectOne")
-    public Subject selectOne(Integer id) {
+    public Subject selectOne(Long id) {
     return subjectService.selectByPrimaryKey(id);
     }
